Fix embedded-svc 0.28 dup, statvfs/sntp removal, OnceLock, get_speed, mi_account u64, transfer Ok wrap + data stub, miwear BLEAddress/poisoned/create_device params, net_http/nvs/logging/astrobox fixes

+++ src/mi_account.rs
@@ -376,7 +376,7 @@ fn next_nonce() -> String {
         .unwrap_or(0);
     // 16 字节混合熵
     let mut bytes = [0u8; 16];
-    let seed = now ^ counter.rotate_left(17);
+    let seed = now ^ (counter as u64).rotate_left(17);
     for i in 0..16 {
         bytes[i] = (seed >> ((i & 7) * 8)) as u8 ^ (counter as u8).wrapping_mul(i as u8);
     }
